+++ luckydonaldUtils/logger/__init__.py
@@ -3,7 +3,7 @@ import logging as _logging
 import sys
 
 __author__ = 'luckydonald'
-__all__ = ["logging", "ColoredFormatter", "ColoredStreamHandler"]
+__all__ = ["logging", "ColoredFormatter", "ColoredStreamHandler", "LevelByNameFilter"]
 
 
 class ColoredFormatter(_logging.Formatter):
@@ -141,6 +141,10 @@ class ColoredFormatter(_logging.Formatter):
 
     def usesTime(self):
         return bool(self.datefmt)
+        # end def
+
+
+# end class
 
 
 class ColoredStreamHandler(_logging.StreamHandler):
@@ -171,17 +175,32 @@ class _LoggingWrapper(object):
         :rtype : int
         :raises KeyError: if the level does not exists.
         """
-        return {
-            "NOTSET": _logging.NOTSET,
-            "DEBUG": _logging.DEBUG,
-            "INFO": _logging.INFO,
-            "SUCCESS": self.SUCCESS,
-            "WARNING": _logging.WARNING,
-            "WARN": _logging.WARN,  # = WARNING
-            "ERROR": _logging.ERROR,
-            "FATAL": _logging.FATAL,  # = CRITICAL
-            "CRITICAL": _logging.CRITICAL,
-        }[level_string]
+        if isinstance(level_string, int):
+            return level_string
+        # end if
+        try:
+            return {
+                # as names:
+                "NOTSET": _logging.NOTSET,
+                "DEBUG": _logging.DEBUG,
+                "INFO": _logging.INFO,
+                "SUCCESS": self.SUCCESS,
+                "WARNING": _logging.WARNING,
+                "WARN": _logging.WARN,  # = WARNING
+                "ERROR": _logging.ERROR,
+                "FATAL": _logging.FATAL,  # = CRITICAL
+                "CRITICAL": _logging.CRITICAL,
+            }[level_string]
+        except KeyError:
+            try:
+                return int(level_string)
+            except ValueError:
+                pass
+            # end try
+            raise  # key not known, and is no integer either.
+            # end try
+
+    # end def
 
     def __call__(self, logger_name):
         """
@@ -191,7 +210,9 @@ class _LoggingWrapper(object):
         """
         return self.getLogger(logger_name)
 
-    def add_colored_handler(self, logger_name=None, stream=None, level=None, date_formatter=None):
+    # end def
+
+    def add_colored_handler(self, logger_name=None, stream=None, level=None, date_formatter=None, filter=None):
         """
         Register a logger handler to colorfull print the messages.
 
@@ -212,29 +233,43 @@ class _LoggingWrapper(object):
                                see https://docs.python.org/library/time.html#time.strftime
         :type    date_formatter: str
 
+        :keyword filter: A filter applied to the handler.
+
         :return: None
         """
         logger = self.getLogger(logger_name)  # __name__
         if stream is None:
             import sys
             stream = sys.stdout
+        # end if
         handler = ColoredStreamHandler(stream=stream, date_formatter=date_formatter)
+        if filter:
+            handler.addFilter(filter)
+        # end if
         logger.addHandler(handler)
         if level:
             logger.setLevel(level)
+        # end if
         return logger
 
-    def test_logger_levels(self):
-        logger = self.getLogger(__name__)
+    # end def
+
+    def test_logger_levels(self, name=__name__, force_all_levels=True):
+        logger = self.getLogger(name)
         logger_level = logger.getEffectiveLevel()
-        logger.setLevel(logging.DEBUG)
+        if force_all_levels:
+            logger.setLevel(logging.DEBUG)
         logger.debug('level debug')
         logger.info('level info')
         logger.success('level success')
         logger.warning('level warning')
         logger.error('level error')
         logger.critical('level critical')
-        logger.setLevel(logger_level)
+        if force_all_levels:
+            logger.setLevel(logger_level)
+            # end if
+
+    # end def
 
     def getLogger(self, name=None):
         """
@@ -280,6 +315,101 @@ class _LoggingWrapper(object):
             pass
         else:
             return getattr(_logging, item)
+            # end def
 
+
+# end class
+
+
+class LevelByNameFilter(object):
+    def __init__(self, root=logging.WARNING, debug=None, info=None, success=None, warning=None, error=None,
+                 critical=None, by_level=None):
+        """
+        A filter where you specify logging levels bound to names (package names, as known from importing)
+
+        :param root: level the root should have to be logged. None to disable.
+
+        :param debug: all loggers which should log debug and above.
+        :param info: all loggers which should log info and above.
+        :param success: all loggers which should log success and above.
+        :param warning: all loggers which should log warning and above.
+        :param error: all loggers which should log error and above.
+        :param critical: all loggers which should log critical and above.
+
+        :param by_level: a dict with levels as a key, and names to log as value.
+                        Example: {10: "__main__", 20: "a.b.c", 30: ["a.b.d", "a.b.e"], logging.WARNING: "a"}
+        """
+        self.mapping = dict()
+        if root:
+            if isinstance(root, str):
+                root = logging.getLoglevelInt(root)
+            assert isinstance(root, int)
+            self.mapping[""] = root
+        # end
+        level = logging.DEBUG
+        self.parse_argument(debug, logging.DEBUG)
+        self.parse_argument(info, logging.INFO)
+        self.parse_argument(success, logging.SUCCESS)
+        self.parse_argument(warning, logging.WARNING)
+        self.parse_argument(error, logging.ERROR)
+        self.parse_argument(critical, logging.CRITICAL)
+
+        if by_level:
+            assert isinstance(by_level, dict)
+            for level, files in by_level.items():
+                self.parse_argument(files, level)
+                # end for
+                # end if
+
+    # end def
+
+    def parse_argument(self, argument, level):
+        if argument:
+            if isinstance(argument, tuple):
+                argument = list(argument)
+            if not isinstance(argument, list):
+                argument = [argument]
+            # end if
+            assert isinstance(argument, list)
+            for part in argument:
+                if isinstance(part, (list, tuple)):
+                    argument.extend(part)
+                elif not isinstance(part, str):
+                    raise TypeError("argument {val!r} is type {type}, should be str.".format(val=part, type=type(part)))
+                elif "," in part:
+                    argument.append(part.split(","))
+                else:
+                    self.mapping[part.strip() + "."] = level
+                    # end if
+                    # end for
+                    # end if
+
+    # end def
+
+    def filter(self, record):
+        if not self.mapping:
+            return False  # allow
+        # end if
+
+        name = record.name + "."
+        mapping_path = ""  # default is "" = root
+
+        for k in self.mapping:
+            if name.startswith(k):
+                if len(mapping_path) < len(k):  # we got a longer path. longer = more specific.
+                    mapping_path = k
+                    # end if
+                    # end if
+        # end for
+
+        if mapping_path in self.mapping:  # e.g. root "" is not specified.
+            level = self.mapping[mapping_path]
+            return record.levelno >= level
+        # end if
+        return False
+        # end def
+
+
+# end class
 
 logging = _LoggingWrapper()